enable SSH ControlPath
A strange issue happend today, the copy of the top content in the
install-server was enough to break its ssh server. It was not
possible anymore to connect.

By enabling the ControlPath, we are still able to reuse the socket and
call the `extract-archive.sh` script to fix the situation.