virtualize.sh: ability to run upgrade
Change the argument:

    virtualize [--hypervisor=bob] version1 versio2 version2

Now it's possible to pass a list of work directory. The first one will
be deployed as usual. virtualize.sh will then run upgrade to the next
version.

During the upgrade process, the /var/lib/edeploy/conf is sourced and
RSERV and RSERV_PORT keys are used.
We had these lines manually.