virtualize.sh: I-S must connect at root@$node
The top/ directory can come with a id_rsa key for the install-server.
In this case, the cloud-init file is used to inject the public key in
the authorized_keys file from the newly deployed nodes.

This is fine as soon as we use PXE/eDeploy. With QCOW2 boot, we lose the
ability to reuse those cloud-init file.

Instead we use an “unsecure” SSH key, we inject the private part in the
install-server and the public part in the authorized_keys of the nodes.
This was working properly with QCOW2, not with eDeploy, because the
public SSH key was not deployed.

With this commit, we preserve the top/root/.ssh/id_rsa and extract the
public SSH that we inject in the nodes with cloud-init.
This commit also ensure the install-server root account can connect to
the root account of every nodes, including the install-server itself.

This commit also pass the public SSH keys to virtualizor.py through the
use of a temporary file (mktemp).

See also the following previous changes:
- 059cc7d7f2dfca228bef7a0b1c6d2a5b03d0a44d
- e7db2fb19b36994f7892c7e4b502e8ec3ceb677a
- 45cd50f3e30816ba6c80aee2b2648a6f40a33384