virtualizor.py: drop all the existing domains
This commit introduce a little behavior changes. So far, --replace
was used to drop a domain when its name was conflicting with a
new domain. The node was removed just before the node creation.

For example, if node A, B and C was running and I need to recreate just
node A and B.
- drop existing node A
- create node A
- drop existing node B
- create node B
The old node C is still running and it's hard to figure out if we are
dealing with the new node B or the old one.

Now, with the `--cleanup` parameter, we drop all the nodes of an
existing prefix at the begining.